Fast math Dreamcast fixes for rmodels.c
GCC SH has a bug in it which causes the compiler to sometimes ICE and
crash upon encountering a bunch of sinf()/cosf() calls within the same
function. This happens quite rarely, but rmodels.c had two functions
which would trigger the issue.

This meant that raylib would not install as a KOS port with -ffast-math
enabled, which should typically always be enabled for performance.

This fix both fixes the compiler ICE issue AND forwards the trig calls
into KOS's trig routines so that they are still using the SH4's FSCA
fast trig approximation instructions, so there's no performance loss
either.